tests: Add virt-install-aarch64-win11
Please note that this is preliminary and some issues remain described
below:

The installation will not start because USB cdrom is not emulated as a
cdrom due to a bug in libvirt:
https://gitlab.com/libvirt/libvirt/-/issues/368

Pass bare QEMU command lines as a workaround.

The display, audio, and TPM do not work because libosinfo cannot tell
if an OS supports devices on a particular architecture:
https://gitlab.com/libosinfo/libosinfo/-/issues/55

In reality, Windows on Arm64 lacks various PCI device drivers and TPM
fails, spamming the Windows system logs. As a workaround, manually
specify ramfb for video and usb for audio and drop the TPM.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>